Getting EAGAIN errors when executing cpanm
I don't grok this. While this fixes the immediate problem, I get extra "OK"
reports on their own line when I do this. I fixed this a different way at first,
just ignoring any undef returned from sysread() and that lead to the
"Configuring" line from cpanm being repeated hundreds of times before outputing
with a newline. I need to figure out what this means and how to fix it the right
way, but this does the trick for the moment.